Rename JellyfishClientSdk to FishjamClientSdk
