+++ .github/workflows/publish_docs.yaml
@@ -12,7 +12,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Build Docs
-        run: mkdir docs && xcodebuild docbuild -scheme JellyfishClientSdk -destination generic/platform=iOS OTHER_DOCC_FLAGS="--transform-for-static-hosting --output-path docs --hosting-base-path ios-client-sdk"
+        run: mkdir docs && xcodebuild docbuild -scheme FishjamClientSdk -destination generic/platform=iOS OTHER_DOCC_FLAGS="--transform-for-static-hosting --output-path docs --hosting-base-path ios-client-sdk"
 
       - name: Install coreutils
         run: brew install coreutils
