+++ FishjamClientDemo/FishjamClientDemo/Assets.xcassets/AccentColor.colorset/Contents.json
