+++ FishjamClientDemo/FishjamClientDemo/Assets.xcassets/AppIcon.appiconset/Contents.json
