+++ FishjamClientDemo/FishjamClientDemo/Assets.xcassets/Contents.json
