+++ FishjamClientDemo/FishjamClientDemo/Preview Content/Preview Assets.xcassets/Contents.json
