+++ FishjamClientDemo/Release.xcconfig
@@ -1,0 +1,1 @@
+jellyfish_url = ws:/$()/192.168.0.164:4000/socket/peer/websocket
