+++ FishjamClientSdk.xcworkspace/contents.xcworkspacedata
