+++ FishjamClientSdk.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
