+++ FishjamClientSdk.xcworkspace/xcshareddata/swiftpm/Package.resolved
