+++ Package.swift
@@ -4,15 +4,15 @@
 import PackageDescription
 
 let package = Package(
-    name: "JellyfishClientSdk",
+    name: "FishjamClientSdk",
     platforms: [
         .iOS(.v13)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "JellyfishClientSdk",
-            targets: ["JellyfishClientSdk"])
+            name: "FishjamClientSdk",
+            targets: ["FishjamClientSdk"])
     ],
     dependencies: [
         .package(name: "WebRTC", url: "https://github.com/webrtc-sdk/Specs.git", .exact("114.5735.08")),
@@ -37,7 +37,7 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "JellyfishClientSdk",
+            name: "FishjamClientSdk",
             dependencies: [
                 "WebRTC", "SwiftProtobuf", "Promises", "SwiftPhoenixClient",
                 .product(name: "Starscream", package: "Starscream"),
@@ -45,6 +45,6 @@ let package = Package(
                 .product(name: "Logging", package: "swift-log"),
             ],
             path: "Sources"),
-        .testTarget(name: "JellyfishClientSdkTests", dependencies: ["JellyfishClientSdk", "Mockingbird"]),
+        .testTarget(name: "FishjamClientSdkTests", dependencies: ["FishjamClientSdk", "Mockingbird"]),
     ]
 )
